Revert "mido: Workaround VoLTE SIM Slot switching issue"
This reverts commit 318d34293cab79a72bab851e1ef2971c3942b514.

Reason: This workaround causes an issue where SIM 2 gets disabled after a reboot, especially if the data SIM is switched.
Because of this, default SIM preferences are also reset and could cause unexpected carrier charges.
If attention is not given after reboot, one could be left with SIM 2 disabled for a longer time and miss calls & SMS.
And if anyone really needs, we can achieve this functionality with a flashable zip!
https://github.com/AliHasan7671/files-store/raw/mido/volte_hack.zip